Refactor code: Use modules
