+++ src/sudoku/field.rs
@@ -1,0 +1,35 @@
+use std::fmt;
+
+#[derive(Debug)]
+pub struct Field {
+    pub row: u8,
+    pub column: u8,
+}
+
+impl fmt::Display for Field {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "({},{})", self.row, self.column)
+    }
+}
+
+impl Field {
+    pub fn new(row: u8, column: u8) -> Field {
+        Field { row, column }
+    }
+}
+
+impl std::cmp::PartialEq for Field {
+    fn eq(&self, other: &Field) -> bool {
+        self.row == other.row && self.column == other.column
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn it_should_format() {
+        assert_eq!(format!("{}", Field::new(1, 2)), "(1,2)");
+    }
+}
