+++ src/sudoku/solver.rs
@@ -1,0 +1,10 @@
+pub enum SolverMethod {
+    Backtracing,
+    Montecarlo,
+}
+
+pub enum EnergyDimension {
+    Row,
+    Column,
+    Parcel,
+}
